Fix compatibility issues with Python 2 in examples.

+++ pynest/examples/hh_psc_alpha.py
@@ -56,7 +56,7 @@ nest.Connect(neuron, sd, syn_spec={'weight': 1.0, 'delay': h})
 
 
 # Simulation loop
-n_data = dcto / float(dcstep)
+n_data = int(dcto / float(dcstep))
 amplitudes = np.zeros(n_data)
 event_freqs = np.zeros(n_data)
 for i, amp in enumerate(range(dcfrom, dcto, dcstep)):
